Move monitor group resource into domain package
